# authentication-service/internal/controller/http/v1/user.go
package v1

import (
	"authentication-service/internal/entity"
	"authentication-service/internal/usecase"
	"authentication-service/pkg/logger"
	"authentication-service/utils"
	"github.com/casbin/casbin/v2"
	"github.com/gin-gonic/gin"
	"net/http"
)

type userRoutes struct {
	t usecase.UserInterface
	l logger.Interface
}

func newUserRoutes(handler *gin.RouterGroup, t usecase.UserInterface, l logger.Interface, csbn *casbin.Enforcer) {
	r := &userRoutes{t, l}

	h := handler.Group("/users")
	{
		//h.GET("/", r.GetTours)
		h.POST("/", r.RegisterUser)
		h.POST("/login", r.LoginUser)
	}
	adm := handler.Group("/admin")
	adm.Use(utils.JWTAuthMiddleware(), utils.CasbinMiddleware(csbn))
	{
		adm.GET("/users", r.GetUsers)
	}
}

func (r *userRoutes) GetUsers(c *gin.Context) {
	users, err := r.t.GetUsers()

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch users"})
		return
	}

	c.JSON(http.StatusOK, &users)
}

func (r *userRoutes) LoginUser(c *gin.Context) {
	var input entity.LoginUserDTO

	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	token, err := r.t.LoginUser(&input)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
	}

	c.JSON(http.StatusOK, gin.H{"token": token})
}

func (r *userRoutes) RegisterUser(c *gin.Context) {
	var createUserDTO entity.CreateUserDTO
	if err := c.ShouldBindJSON(&createUserDTO); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	hashedPassword, err := utils.HashPassword(createUserDTO.Password)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error hashing password"})
		return
	}

	user := entity.User{
		Username: createUserDTO.Username,
		Email:    createUserDTO.Email,
		Password: hashedPassword,
		Role:     "user",
	}

	createdUser, err := r.t.RegisterUser(&user)

	c.JSON(http.StatusCreated, gin.H{"message": "User registered successfully", "User": createdUser})
}